Slight efficiency gain.
Removes a declaration from the draw loop so that the variable is declared once and not like every two frames. This variable should change after every draw call because it is overwritten with new values. Before it would've had to redeclare a already declared variable and then write a new value to that.